requirements.txt


senetence-transformers==2.2.2 
flask
pypdf2 
python-dotenv 
pinecone[grpc] 
langchain-pinecone 
langchain_community  
langchain_openai
langchain_experimental
-e .